tank service singleton central point of access
TankService is main file which is going to be attached to an empty tankspawner gameobject.

System implemented in the style of
       TankService -> TankController  -> TankView

TankController can access(read-only) both TankView and TankModel but not other way round.
TankModel stores data values for tank which can be modified from the TankService handler.
TankService is the point of access for creating tanks of different specs.
TankView handles movement by using rigidbody.