tankScriptableObject and tankScriptableObjectList
tankScriptableObject  will be holding all the data required for the tankmodel to get its data values from. Currently created 3 types red, green and blue.

tankScriptableObjectList: will be holding all the tankScriptableObject in a array.
keep in mind the tankScriptableObjectList will always be one shorter than the tankType enums created, because currently not filling up  the tankScriptableObjectList with first element as TankType.None